refactor: create mud_string.cpp with MUD argument parsing functions (#3007)
Move MUD-specific argument parsing from interpreter.cpp to new
src/utils/mud_string.cpp:
- dir_fill/reserved arrays
- fill_word, reserved_word
- one_argument (2 overloads + template)
- any_one_arg (2 overloads + template)
- two_arguments, three_arguments (templates in .h)
- SplitArgument (3 overloads)
- half_chop

All declarations in mud_string.h with Russian comments.
utils.h includes mud_string.h for backward compatibility.
